Fix a few small warnings related to format strings
* In 2 cases, %u was being used with NSUInteger, which is an unsigned
  long on 64-bit systems. Cast the value to (unsigned int), since the
  expected values only go up to 6.
* nil was used erroneously at the end of an argument list for one
  format string.

+++ Source/MMGenerator.m
@@ -83,7 +83,7 @@ static NSString * __HTMLStartTagForElement(MMElement *anElement)
     switch (anElement.type)
     {
         case MMElementTypeHeader:
-            return [NSString stringWithFormat:@"<h%u>", anElement.level];
+            return [NSString stringWithFormat:@"<h%u>", (unsigned int)anElement.level];
         case MMElementTypeParagraph:
             return @"<p>";
         case MMElementTypeBulletedList:
@@ -140,7 +140,7 @@ static NSString * __HTMLEndTagForElement(MMElement *anElement)
     switch (anElement.type)
     {
         case MMElementTypeHeader:
-            return [NSString stringWithFormat:@"</h%u>\n", anElement.level];
+            return [NSString stringWithFormat:@"</h%u>\n", (unsigned int)anElement.level];
         case MMElementTypeParagraph:
             return @"</p>\n";
         case MMElementTypeBulletedList:
